redoing refactors after main commit
