+++ webui/src/routes/dashboard.tsx
@@ -7,6 +7,8 @@ export const Route = createFileRoute("/dashboard")({
 });
 
 function LayoutComponent() {
+  // TODO: user needs to be logged in to access the dashboard if user is not logged in, redirect to the login page
+
   const queryClient = new QueryClient();
   return (
     <QueryClientProvider client={queryClient}>
