+++ webui/src/services/repository.ts
@@ -1,0 +1,24 @@
+import type { Status } from "../types/API";
+import type { Repository } from "../types/Repository";
+import { fetchWithRedirect } from "./common";
+
+// Fetch the repository by ID with owner validation
+export async function getRepositoryByIDWithOwnerValidation(repoID: number, repoFullName: string): Promise<Status> {
+  const r = await fetchWithRedirect(`/api/dashboard/repository/${repoID}?repo=${repoFullName}`);
+  return { id: repoID, status: r.status };
+}
+
+// Fetch the user's repositories
+export async function getUserRepositories(): Promise<Repository[]> {
+  const r = await fetchWithRedirect("/api/dashboard/repositories");
+  return (await r.json()) as Repository[];
+}
+
+export async function createNewRepository(params: URLSearchParams): Promise<Response> {
+  return fetchWithRedirect(
+    `/api/dashboard/repository/new?${params.toString()}`,
+    {
+      method: "POST",
+    },
+  );
+}
