+++ webui/src/types/User.d.ts
@@ -1,0 +1,9 @@
+export type User = {
+  id: number;
+  login: string;
+  email: string;
+  avatar_url: string;
+  name: string;
+  type: string;
+};
+
